Move shopify dev renderer into own folder, begin to componentize
