+++ scripts/typedoc/types.ts
@@ -1,3 +1,14 @@
+export interface Packages {
+  [key: string]: string;
+}
+
+export interface Paths {
+  inputRoot: string;
+  outputRoot: string;
+  packages: Packages;
+  shopifyDevUrl: string;
+}
+
 export interface Tag {
   name: string;
   content: string;
